More refactor and fmt
